+++ src/backendDocker.go
@@ -44,6 +44,10 @@ func (d *backendDocker) Arch() TypeArch {
 	return TypeArchAmd
 }
 
+func (d *backendDocker) DeleteSecurityGroups(vpc string) error {
+	return nil
+}
+
 func (d *backendDocker) ClusterList() ([]string, error) {
 	out, err := exec.Command("docker", "container", "list", "-a", "--format", "{{json .Names}}").CombinedOutput()
 	if err != nil {