+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-#### 4.4.5
+#### 4.5.0
 * fix bug with old Aerospike version download, requiring username/password
 * add notes at the end of AMS and exporter installation, referencing one-another to avoid confusion
 * allow for aerolab to auto-create security groups in aws, as well as discover subnet IDs based on AZ for the default VPC